feat: refactor deserialization to use uint instead of ulong for length; improve data handling in multiple components

+++ Libraries/Flakkari4Unity/Source/Components/Common/Tag.cs
@@ -9,8 +9,8 @@ namespace Flakkari4Unity.ECS.Components.Common
 
         internal void Deserialize(byte[] data, ref int i)
         {
-            ulong length = BitConverter.ToUInt64(data, i);
-            i += sizeof(ulong);
+            uint length = BitConverter.ToUInt32(data, i);
+            i += sizeof(uint);
             name = System.Text.Encoding.UTF8.GetString(data, i, (int)length);
             i += (int)length;
         }
